+++ dash-abci/Cargo.toml
@@ -6,4 +6,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+rust_decimal = "1.2.5"
+rust_decimal_macros = "1.25"
 rs-drive = { path = "../drive" }
+
+[dependencies.dpp]
+git = "https://github.com/dashevo/rust-dpp"
+rev = "ee4daa16a7aaa8d46f075f830af9e63b10f54628"
