+++ dash-abci/src/common/helpers/fee_pools.rs
@@ -1,16 +1,14 @@
 use std::collections::BTreeMap;
 use std::ops::Range;
 
-use ciborium::value::Value;
 use dpp::{
     identifier::Identifier,
     identity::{Identity, IdentityPublicKey, KeyType},
 };
-use grovedb::TransactionArg;
+use rs_drive::grovedb::TransactionArg;
 use rs_drive::drive::batch::GroveDbOpBatch;
 use rs_drive::drive::Drive;
 use rs_drive::fee_pools::epochs::Epoch;
-use rs_drive::query::TransactionArg;
 
 pub fn get_storage_credits_for_distribution_for_epochs_in_range(
     drive: &Drive,
