+++ dash-abci/src/execution/epoch_change/mod.rs
@@ -1,1 +1,2 @@
 mod distribute_storage_pool;
+pub mod epoch;
