+++ drive/Cargo.toml
@@ -26,8 +26,6 @@ thiserror = "1.0.30"
 moka = "0.8.1"
 nohash-hasher = "0.2.0"
 chrono = "0.4.0"
-rust_decimal = "1.2.5"
-rust_decimal_macros = "1.25"
 
 [dependencies.grovedb]
 git = "https://github.com/dashevo/grovedb"
