+++ drive/src/drive/fee_pools/mod.rs
@@ -1,7 +1,6 @@
-use crate::drive::fee_pools::constants::KEY_STORAGE_FEE_POOL;
 use crate::drive::RootTree;
+use crate::fee_pools::epochs_root_tree_key_constants::KEY_STORAGE_FEE_POOL;
 
-pub mod constants;
 pub mod epochs;
 pub mod storage_fee_distribution_pool;
 
@@ -22,7 +21,6 @@ pub(crate) fn aggregate_storage_fees_distribution_pool_vec_path() -> Vec<Vec<u8>
 
 #[cfg(test)]
 mod tests {
-    use crate::common::tests::helpers::setup::setup_drive;
     use crate::common::tests::helpers::setup::setup_drive_with_initial_state_structure;
     use crate::drive::batch::GroveDbOpBatch;
     use crate::error;
