+++ drive/src/drive/fee_pools/storage_fee_distribution_pool.rs
@@ -1,14 +1,10 @@
-use crate::drive::fee_pools::{
-    aggregate_storage_fees_distribution_pool_path,
-    aggregate_storage_fees_distribution_pool_vec_path,
-};
+use crate::drive::fee_pools::aggregate_storage_fees_distribution_pool_path;
 use crate::drive::Drive;
 use grovedb::{Element, TransactionArg};
 
 use crate::error::fee::FeeError;
 use crate::error::Error;
-
-use super::constants;
+use crate::fee_pools::epochs_root_tree_key_constants::KEY_STORAGE_FEE_POOL;
 
 impl Drive {
     pub fn get_aggregate_storage_fees_in_current_distribution_pool(
@@ -19,7 +15,7 @@ impl Drive {
             .grove
             .get(
                 aggregate_storage_fees_distribution_pool_path(),
-                constants::KEY_STORAGE_FEE_POOL.as_slice(),
+                KEY_STORAGE_FEE_POOL.as_slice(),
                 transaction,
             )
             .unwrap()
