+++ drive/src/fee/mod.rs
@@ -5,7 +5,6 @@ use crate::error::Error;
 use crate::fee::op::{BaseOp, DriveCost, DriveOperation};
 
 pub mod default_costs;
-pub mod epoch;
 pub mod fees_aggregate;
 pub mod op;
