+++ drive/src/fee_pools/epochs/mod.rs
@@ -1,6 +1,6 @@
 pub mod operations_factory;
 pub mod paths;
-pub mod tree_key_constants;
+pub mod epoch_key_constants;
 
 use serde::{Deserialize, Serialize};
