+++ drive/src/fee_pools/epochs_root_tree_key_constants.rs
@@ -1,0 +1,1 @@
+pub const KEY_STORAGE_FEE_POOL: &[u8; 1] = b"s";
