+++ drive/src/fee_pools/mod.rs
@@ -1,12 +1,13 @@
 use crate::drive::batch::GroveDbOpBatch;
-use crate::drive::fee_pools::constants::KEY_STORAGE_FEE_POOL;
 use crate::drive::fee_pools::fee_pool_vec_path;
 use crate::fee_pools::epochs::Epoch;
 use grovedb::batch::GroveDbOp;
 use grovedb::batch::Op::Insert;
 use grovedb::Element;
+use crate::fee_pools::epochs_root_tree_key_constants::KEY_STORAGE_FEE_POOL;
 
 pub mod epochs;
+pub mod epochs_root_tree_key_constants;
 
 pub fn add_create_fee_pool_trees_operations(batch: &mut GroveDbOpBatch) {
     // Update storage credit pool
