+++ drive/src/query/mod.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 use std::ops::BitXor;
 
 use ciborium::value::Value;
-pub use grovedb::{
+use grovedb::{
     Element, Error as GroveError, GroveDb, PathQuery, Query, QueryItem, SizedQuery, TransactionArg,
 };
 use indexmap::IndexMap;
